+++ .dotnet.azure/src/Custom/Chat/AzureChatClient.cs
@@ -21,17 +21,17 @@ internal partial class AzureChatClient : ChatClient
     private readonly Uri _endpoint;
     private readonly string _apiVersion;
 
-    internal AzureChatClient(
-        ClientPipeline pipeline,
-        string deploymentName,
-        Uri endpoint,
-        AzureOpenAIClientOptions options)
-            : base(pipeline, model: deploymentName, endpoint, options)
+    internal AzureChatClient(ClientPipeline pipeline, string deploymentName, Uri endpoint, AzureOpenAIClientOptions options)
+        : base(pipeline, model: deploymentName, new OpenAIClientOptions() { Endpoint = endpoint })
     {
+        Argument.AssertNotNull(pipeline, nameof(pipeline));
+        Argument.AssertNotNullOrEmpty(deploymentName, nameof(deploymentName));
+        Argument.AssertNotNull(endpoint, nameof(endpoint));
         options ??= new();
+
         _deploymentName = deploymentName;
-        _endpoint = endpoint;
         _apiVersion = options.Version;
+        _endpoint = endpoint;
     }
 
     protected AzureChatClient()